Adding frontend app and changing names accordingly
