--- src/backend/interface/controller/app_controller.go
@@ -1,5 +1,0 @@
-package controller
-
-type AppController interface {
-	PokemonController
-}
