--- src/backend/interface/controller/context.go
@@ -1,6 +1,0 @@
-package controller
-
-type Context interface {
-	JSON(code int, i interface{}) error
-	Bind(i interface{}) error
-}
