--- src/backend/interface/logger.go
@@ -1,6 +1,0 @@
-package interfaces
-
-type Logger interface {
-	LogError(string, ...interface{})
-	LogAccess(string, ...interface{})
-}
