--- src/backend/usecase/presenter/pokemon_presenter.go
@@ -1,7 +1,0 @@
-package presenter
-
-import "github.com/GerardoHP/ondemand-go-bootcamp/domain/model"
-
-type PokemonPresenter interface {
-	ResponsePresenter(p []*model.Pokemon) []*model.Pokemon
-}
