--- src/backend/usecase/repository/pokemon_repository.go
@@ -1,7 +1,0 @@
-package repository
-
-import "github.com/GerardoHP/ondemand-go-bootcamp/domain/model"
-
-type PokemonRepositoty interface {
-	FindAll(p []*model.Pokemon) ([]*model.Pokemon, error)
-}
