+++ src/pokemon-service/interface/controller/app_controller.go
@@ -1,0 +1,5 @@
+package controller
+
+type AppController interface {
+	PokemonController
+}
