+++ src/pokemon-service/interface/controller/context.go
@@ -1,0 +1,6 @@
+package controller
+
+type Context interface {
+	JSON(code int, i interface{}) error
+	Bind(i interface{}) error
+}
