+++ src/pokemon-service/interface/logger.go
@@ -1,0 +1,6 @@
+package interfaces
+
+type Logger interface {
+	LogError(string, ...interface{})
+	LogAccess(string, ...interface{})
+}
