+++ src/pokemon-service/interface/presenter/pokemon_presenter.go
@@ -1,0 +1,26 @@
+package presenter
+
+import (
+	"strings"
+
+	"github.com/GerardoHP/ondemand-go-bootcamp/domain/model"
+)
+
+type pokemonPresenter struct {
+}
+
+type PokemonPresenter interface {
+	ResponsePresenter(pk []*model.Pokemon) []*model.Pokemon
+}
+
+func NewPokemonPresenter() PokemonPresenter {
+	return &pokemonPresenter{}
+}
+
+func (pp *pokemonPresenter) ResponsePresenter(pk []*model.Pokemon) []*model.Pokemon {
+	for _, p := range pk {
+		p.Name = strings.ToUpper(p.Name)
+	}
+
+	return pk
+}
