+++ src/pokemon-service/pokemons.csv
@@ -1,0 +1,20 @@
+1, bulbasaur, https://pokeapi.co/api/v2/pokemon/1
+2, ivysaur, https://pokeapi.co/api/v2/pokemon/2
+3, venusaur, https://pokeapi.co/api/v2/pokemon/3
+4, charmander, https://pokeapi.co/api/v2/pokemon/4
+5, charmeleon, https://pokeapi.co/api/v2/pokemon/5
+6, charizard, https://pokeapi.co/api/v2/pokemon/6
+7, squirtle, https://pokeapi.co/api/v2/pokemon/7
+8, wartortle, https://pokeapi.co/api/v2/pokemon/8
+9, blastoise, https://pokeapi.co/api/v2/pokemon/9
+10, caterpie, https://pokeapi.co/api/v2/pokemon/10
+11, metapod, https://pokeapi.co/api/v2/pokemon/11
+12, butterfree, https://pokeapi.co/api/v2/pokemon/12
+13, weedle, https://pokeapi.co/api/v2/pokemon/13
+14, kakuna, https://pokeapi.co/api/v2/pokemon/14
+15, beedrill, https://pokeapi.co/api/v2/pokemon/15
+16, pidgey, https://pokeapi.co/api/v2/pokemon/16
+17, pidgeotto, https://pokeapi.co/api/v2/pokemon/17
+18, pidgeot, https://pokeapi.co/api/v2/pokemon/18
+19, rattata, https://pokeapi.co/api/v2/pokemon/19
+20, raticate, https://pokeapi.co/api/v2/pokemon/20
