+++ src/pokemon-service/usecase/repository/pokemon_repository.go
@@ -1,0 +1,7 @@
+package repository
+
+import "github.com/GerardoHP/ondemand-go-bootcamp/domain/model"
+
+type PokemonRepositoty interface {
+	FindAll(p []*model.Pokemon) ([]*model.Pokemon, error)
+}
